+++ assets/script/components/pools/TransactionPoolByPerson.tsx
@@ -13,6 +13,8 @@ import { FormType } from '../form/FormHelper';
 import TransactionCRUD from '../form/CRUD/TransactionCRUD';
 import GraphqlService from '../../helpers/GraphqlService';
 import { searchFilter, searchSelector } from '../../helpers/SearchHelper';
+import { amountRender } from '../../helpers/AmountInput';
+import { dateRender } from '../../helpers/DateHelper';
 
 interface TransactionPoolState {
   searchAttribute: string | Array<string> | null,
@@ -117,8 +119,8 @@ function TransactionPoolByPerson() {
   const expandedRowRender = (record:PersonTransactions) => {
     const columns: ColumnsType<Transaction> = [
       { title: 'Titel', dataIndex: 'getTitle', key: 'title' },
-      { title: 'Datum', dataIndex: 'getDate', key: 'date' },
-      { title: 'Bedrag', dataIndex: 'amount', key: 'amount' },
+      { title: 'Datum', key: 'date', render: (_, { getDate }) => dateRender(getDate) },
+      { title: 'Bedrag', key: 'amount', render: (_, { amount }) => amountRender(amount) },
       { title: 'Bankaccount', dataIndex: ['getTransactionGroup', 'getBankAccount', 'name'], key: 'account' },
       {
         title: 'Status',
